src/requirements.txt
absl-py==1.4.0
accelerate==0.23.0
ai-tools==0.3.9
aiohttp==3.8.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-beam==2.54.0
asciitree==0.3.3
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
atomics==1.0.2
attrdict==2.0.1
attrs==23.2.0
av==11.0.0
backoff==2.2.1
bcrypt==4.2.0
beartype==0.15.0
blis==0.7.11
blobfile==2.1.1
blurhash==1.1.4
boost==0.1
braceexpand==0.1.7
build==1.2.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.5.5
ci-info==0.3.0
clean-fid==0.1.35
click==8.1.7
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
clip-anytorch==2.5.2
cloudpathlib==0.16.0
cmake==3.27.5
CoCa-pytorch==0.0.12
cognitive-face==1.5.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
crcmod==1.7
cryptography==43.0.0
cymem==2.0.8
Cython==3.0.8
dalle2-pytorch==1.15.1
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.21.4
distro==1.9.0
dlib-binary==19.24.1
dnspython==2.6.0
docker-pycreds==0.4.0
docopt==0.6.2
dominate==2.8.0
easydict==1.10
einops==0.6.1
einops-exts==0.0.4
ema-pytorch==0.2.3
embedding-reader==1.5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.13
fastapi==0.99.1
fastavro==1.9.4
fastcore==1.5.29
fasteners==0.19
fastlogging==1.0.0
fastrlock==0.8.2
ffmpeg==1.4
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.1.1
func-timeout==4.3.5
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.34.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gql==3.4.1
graphql-core==3.2.3
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.66.0
gym-notices==0.0.8
Gymnasium==0.26.3
h11==0.14.0
halo==0.0.31
hdfs==2.7.3
himalaya==0.4.2
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.3.2
icecream==2.1.3
idna==3.7
imhist==0.0.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
imWatermark==0.0.2
invisible-watermark==0.2.0
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
jieba==0.42.1
jiter==0.5.0
Js2Py==0.74
jsonmerge==1.9.2
jsonpatch==1.33
jsonpath==0.82.2
jsonpickle==3.0.2
jsonpointer==3.0.0
k-diffusion==0.0.16
kornia==0.7.0
kubernetes==30.1.0
langchain==0.1.19
langchain-anthropic==0.1.23
langchain-community==0.0.38
langchain-core==0.2.36
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langcodes==3.3.0
langgraph==0.2.14
langgraph-checkpoint==1.0.6
langsmith==0.1.80
libclang==16.0.6
lightning-utilities==0.9.0
log-symbols==0.0.14
looseversion==1.3.0
lpips==0.1.4
markdown-it-py==3.0.0
marshmallow==3.21.3
Mastodon.py==1.8.1
mat4py==0.5.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.1.0
monai==1.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
natsort==8.4.0
nibabel==5.1.0
nilearn==0.10.1
nipype==1.8.6
numcodecs==0.12.1
numpy==1.26.4
oauthlib==3.2.2
objsize==0.7.0
omegaconf==2.3.0
onnxruntime==1.18.0
open-clip-torch==2.20.0
openai==1.34.0
opencv-python==4.8.0.74
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt-einsum==3.3.0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
pathtools==0.1.2
plotly-express==0.4.1
posthog==3.5.0
preshed==3.0.9
prettytable==3.8.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==5.27.1
prov==2.0.0
pulsar-client==3.5.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-mini-racer==0.6.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.19.0
pydantic==2.8.2
pydantic_core==2.20.1
pyDeprecate==0.3.2
pydot==1.4.2
pyecharts==2.0.4
pyemd==1.0.0
PyExecJS==1.5.1
pygame==2.5.2
PyGithub==2.4.0
pyglet==1.5.27
Pygments==2.16.1
pyjsparser==2.7.1
PyJWT==2.9.0
pymongo==4.6.1
pymunk==6.6.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyNaCl==1.5.0
pynput==1.7.6
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyrealsense2==2.54.2.5684
pytextrank==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==1.7.7
pytorch-warmup==0.1.1
pytorchtools==0.0.2
pytz==2024.1
PyYAML==6.0.1
qstock==1.3.6
ray==2.4.0
rdflib==6.3.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.5.3
rotary-embedding-torch==0.3.0
rsa==4.9
scikit-video==1.1.11
scripts==2.0
sentence-transformers==2.7.0
sentencepiece==0.1.99
sentry-sdk==1.27.0
setproctitle==1.3.2
shapely==2.0.2
shellingham==1.5.3
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
spnav @ https://github.com/cheng-chi/spnav/archive/c1c938ebe3cc542db4685e0d13850ff1abfdb943.tar.gz
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
stua==0.2
sympy==1.12.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==8.4.1
tensorboard-logger==0.1.0
tensorboardX==2.6.1
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.2.2
thop==0.1.1.post2209072238
tiktoken==0.7.0
timm==0.9.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.1+cu118
torch-fidelity==0.3.0
torch-geometric==2.3.1
torchaudio==2.2.1+cu118
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchsummary==1.5.1
torchvision==0.17.1+cu118
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
trampoline==0.1.2
transform==1.2.0
transform-tools==1.0.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
update-checker==0.18.0
ur-rtde==1.5.7
urllib3==2.2.2
uvicorn==0.30.1
validators==0.22.0
vector-quantize-pytorch==1.7.1
visdom==0.2.4
wandb==0.15.5
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.48
websocket-client==1.8.0
websockets==12.0
whl==0.0.4
wrapt==1.16.0
x-clip==0.12.9
XlsxWriter==3.1.9
xxhash==3.3.0
yarl==1.9.4
zarr==2.16.1
zipp==3.20.0
zprint==0.0.11
